Move benchmarks to its own folder/executable
